Move all phone number queries to neo4j module
