+++ packages/server/customer-os-neo4j-repository/repository/phone_number_write_repository.go
@@ -20,10 +20,22 @@ type PhoneNumberCreateFields struct {
 	UpdatedAt      time.Time    `json:"updatedAt"`
 }
 
+type PhoneNumberValidateFields struct {
+	E164          string    `json:"e164"`
+	CountryCodeA2 string    `json:"countryCodeA2"`
+	ValidatedAt   time.Time `json:"validatedAt"`
+	Source        string    `json:"source"`
+	AppSource     string    `json:"appSource"`
+}
+
 type PhoneNumberWriteRepository interface {
 	CreatePhoneNumber(ctx context.Context, tenant, phoneNumberId string, data PhoneNumberCreateFields) error
 	UpdatePhoneNumber(ctx context.Context, tenant, phoneNumberId, source string, updatedAt time.Time) error
 	FailPhoneNumberValidation(ctx context.Context, tenant, phoneNumberId, validationError string, validatedAt time.Time) error
+	PhoneNumberValidated(ctx context.Context, tenant, phoneNumberId string, data PhoneNumberValidateFields) error
+	LinkWithContact(ctx context.Context, tenant, contactId, phoneNumberId, label string, primary bool, updatedAt time.Time) error
+	LinkWithOrganization(ctx context.Context, tenant, organizationId, phoneNumberId, label string, primary bool, updatedAt time.Time) error
+	LinkWithUser(ctx context.Context, tenant, userId, phoneNumberId, label string, primary bool, updatedAt time.Time) error
 }
 
 type phoneNumberWriteRepository struct {
@@ -121,6 +133,153 @@ func (r *phoneNumberWriteRepository) FailPhoneNumberValidation(ctx context.Conte
 		"tenant":          tenant,
 		"validationError": validationError,
 		"validatedAt":     validatedAt,
+	}
+	span.LogFields(log.String("cypher", cypher))
+	tracing.LogObjectAsJson(span, "params", params)
+
+	err := utils.ExecuteWriteQuery(ctx, *r.driver, cypher, params)
+	if err != nil {
+		tracing.TraceErr(span, err)
+	}
+	return err
+}
+
+func (r *phoneNumberWriteRepository) PhoneNumberValidated(ctx context.Context, tenant, phoneNumberId string, data PhoneNumberValidateFields) error {
+	span, ctx := opentracing.StartSpanFromContext(ctx, "PhoneNumberWriteRepository.PhoneNumberValidated")
+	defer span.Finish()
+	tracing.SetNeo4jRepositorySpanTags(span, tenant)
+	span.SetTag(tracing.SpanTagEntityId, phoneNumberId)
+	tracing.LogObjectAsJson(span, "data", data)
+
+	cypher := fmt.Sprintf(`MATCH (t:Tenant {name:$tenant})<-[:PHONE_NUMBER_BELONGS_TO_TENANT]-(p:PhoneNumber {id:$id})
+				WHERE p:PhoneNumber_%s
+		 		SET p.validationError = $validationError,
+					p.e164 = $e164,
+		     		p.validated = true,
+					p.updatedAt = $validatedAt
+				WITH p
+				WHERE $countryCodeA2 <> ''
+				WITH p
+				CALL {
+					WITH p
+    				OPTIONAL MATCH (p)-[r:LINKED_TO]->(oldCountry:Country)
+    				WHERE oldCountry.codeA2 <> $countryCodeA2
+    				DELETE r
+				}
+				MERGE (c:Country {codeA2: $countryCodeA2})
+					ON CREATE SET 	c.createdAt = $now, 
+									c.updatedAt = $now, 
+									c.appSource = $appSource,
+									c.source = $source,
+									c.sourceOfTruth = $source
+				MERGE (p)-[:LINKED_TO]->(c)
+				`, tenant)
+	params := map[string]any{
+		"id":              phoneNumberId,
+		"tenant":          tenant,
+		"validationError": "",
+		"e164":            data.E164,
+		"validatedAt":     data.ValidatedAt,
+		"countryCodeA2":   data.CountryCodeA2,
+		"now":             utils.Now(),
+		"appSource":       data.AppSource,
+		"source":          data.Source,
+	}
+	span.LogFields(log.String("cypher", cypher))
+	tracing.LogObjectAsJson(span, "params", params)
+
+	err := utils.ExecuteWriteQuery(ctx, *r.driver, cypher, params)
+	if err != nil {
+		tracing.TraceErr(span, err)
+	}
+	return err
+}
+
+func (r *phoneNumberWriteRepository) LinkWithContact(ctx context.Context, tenant, contactId, phoneNumberId, label string, primary bool, updatedAt time.Time) error {
+	span, ctx := opentracing.StartSpanFromContext(ctx, "PhoneNumberWriteRepository.LinkWithContact")
+	defer span.Finish()
+	tracing.SetNeo4jRepositorySpanTags(span, tenant)
+	span.SetTag(tracing.SpanTagEntityId, phoneNumberId)
+
+	cypher := `
+		MATCH (t:Tenant {name:$tenant})<-[:CONTACT_BELONGS_TO_TENANT]-(c:Contact {id:$contactId}),
+				(t)<-[:PHONE_NUMBER_BELONGS_TO_TENANT]-(p:PhoneNumber {id:$phoneNumberId})
+		MERGE (c)-[rel:HAS]->(p)
+		SET	rel.primary = $primary,
+			rel.label = $label,	
+			c.updatedAt = $updatedAt,
+			rel.syncedWithEventStore = true`
+	params := map[string]any{
+		"tenant":        tenant,
+		"contactId":     contactId,
+		"phoneNumberId": phoneNumberId,
+		"label":         label,
+		"primary":       primary,
+		"updatedAt":     updatedAt,
+	}
+	span.LogFields(log.String("cypher", cypher))
+	tracing.LogObjectAsJson(span, "params", params)
+
+	err := utils.ExecuteWriteQuery(ctx, *r.driver, cypher, params)
+	if err != nil {
+		tracing.TraceErr(span, err)
+	}
+	return err
+}
+
+func (r *phoneNumberWriteRepository) LinkWithOrganization(ctx context.Context, tenant, organizationId, phoneNumberId, label string, primary bool, updatedAt time.Time) error {
+	span, ctx := opentracing.StartSpanFromContext(ctx, "PhoneNumberWriteRepository.LinkWithOrganization")
+	defer span.Finish()
+	tracing.SetNeo4jRepositorySpanTags(span, tenant)
+	span.SetTag(tracing.SpanTagEntityId, phoneNumberId)
+
+	cypher := `
+		MATCH (t:Tenant {name:$tenant})<-[:ORGANIZATION_BELONGS_TO_TENANT]-(org:Organization {id:$organizationId}),
+				(t)<-[:PHONE_NUMBER_BELONGS_TO_TENANT]-(p:PhoneNumber {id:$phoneNumberId})
+		MERGE (org)-[rel:HAS]->(p)
+		SET	rel.primary = $primary,
+			rel.label = $label,	
+			org.updatedAt = $updatedAt,
+			rel.syncedWithEventStore = true`
+	params := map[string]any{
+		"tenant":         tenant,
+		"organizationId": organizationId,
+		"phoneNumberId":  phoneNumberId,
+		"label":          label,
+		"primary":        primary,
+		"updatedAt":      updatedAt,
+	}
+	span.LogFields(log.String("cypher", cypher))
+	tracing.LogObjectAsJson(span, "params", params)
+
+	err := utils.ExecuteWriteQuery(ctx, *r.driver, cypher, params)
+	if err != nil {
+		tracing.TraceErr(span, err)
+	}
+	return err
+}
+
+func (r *phoneNumberWriteRepository) LinkWithUser(ctx context.Context, tenant, userId, phoneNumberId, label string, primary bool, updatedAt time.Time) error {
+	span, ctx := opentracing.StartSpanFromContext(ctx, "PhoneNumberWriteRepository.LinkWithUser")
+	defer span.Finish()
+	tracing.SetNeo4jRepositorySpanTags(span, tenant)
+	span.SetTag(tracing.SpanTagEntityId, phoneNumberId)
+
+	cypher := `
+		MATCH (t:Tenant {name:$tenant})<-[:USER_BELONGS_TO_TENANT]-(u:User {id:$userId}),
+				(t)<-[:PHONE_NUMBER_BELONGS_TO_TENANT]-(p:PhoneNumber {id:$phoneNumberId})
+		MERGE (u)-[rel:HAS]->(p)
+		SET	rel.primary = $primary,
+			rel.label = $label,	
+			u.updatedAt = $updatedAt,
+			rel.syncedWithEventStore = true`
+	params := map[string]any{
+		"tenant":        tenant,
+		"userId":        userId,
+		"phoneNumberId": phoneNumberId,
+		"label":         label,
+		"primary":       primary,
+		"updatedAt":     updatedAt,
 	}
 	span.LogFields(log.String("cypher", cypher))
 	tracing.LogObjectAsJson(span, "params", params)
