+++ packages/server/events-processing-platform/repository/repositories.go
@@ -4,7 +4,6 @@ import (
 	"github.com/neo4j/neo4j-go-driver/v5/neo4j"
 	cmn_repository "github.com/openline-ai/openline-customer-os/packages/server/customer-os-common-module/repository"
 	neo4jrepository "github.com/openline-ai/openline-customer-os/packages/server/customer-os-neo4j-repository/repository"
-	"github.com/openline-ai/openline-customer-os/packages/server/events-processing-platform/logger"
 	repository "github.com/openline-ai/openline-customer-os/packages/server/events-processing-platform/repository/postgres"
 	"github.com/openline-ai/openline-customer-os/packages/server/events-processing-platform/repository/postgres/entity"
 	"gorm.io/gorm"
@@ -28,10 +27,9 @@ type Repositories struct {
 	LocationRepository         LocationRepository
 	OpportunityRepository      OpportunityRepository
 	OrganizationRepository     OrganizationRepository
-	PhoneNumberRepository      PhoneNumberRepository
 }
 
-func InitRepos(driver *neo4j.DriverWithContext, neo4jDatabase string, gormDb *gorm.DB, log logger.Logger) *Repositories {
+func InitRepos(driver *neo4j.DriverWithContext, neo4jDatabase string, gormDb *gorm.DB) *Repositories {
 	repositories := Repositories{
 		Drivers: Drivers{
 			Neo4jDriver: driver,
@@ -39,7 +37,6 @@ func InitRepos(driver *neo4j.DriverWithContext, neo4jDatabase string, gormDb *go
 		Neo4jRepositories:          neo4jrepository.InitNeo4jRepositories(driver, neo4jDatabase),
 		CommonRepositories:         cmn_repository.InitRepositories(gormDb, driver),
 		CustomerOsIdsRepository:    repository.NewCustomerOsIdsRepository(gormDb),
-		PhoneNumberRepository:      NewPhoneNumberRepository(driver),
 		EmailRepository:            NewEmailRepository(driver),
 		OrganizationRepository:     NewOrganizationRepository(driver, neo4jDatabase),
 		LocationRepository:         NewLocationRepository(driver),
